+++ src/skiasharp/LiveChartsCore.SkiaSharp.WPF/CartesianChart.cs
@@ -21,54 +21,21 @@
 // SOFTWARE.
 
 using LiveChartsCore.Context;
-using LiveChartsCore.Drawing;
-using LiveChartsCore.Rx;
 using LiveChartsCore.SkiaSharpView.Drawing;
-using SkiaSharp.Views.WPF;
-using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Media;
 
 namespace LiveChartsCore.SkiaSharpView.WPF
 {
-    public class CartesianChart : Control, ICartesianChartView<SkiaDrawingContext>
+    public class CartesianChart : Chart, ICartesianChartView<SkiaDrawingContext>
     {
-        protected CartesianChart<SkiaDrawingContext> core;
-        protected NaturalGeometriesCanvas canvas;
-        protected IChartLegend<SkiaDrawingContext> legend;
-        protected IChartTooltip<SkiaDrawingContext> tooltip;
-        private readonly ActionThrottler mouseMoveThrottler;
-        private PointF mousePosition = new PointF();
-
         static CartesianChart()
         {
             DefaultStyleKeyProperty.OverrideMetadata(typeof(CartesianChart), new FrameworkPropertyMetadata(typeof(CartesianChart)));
         }
 
-        public CartesianChart()
-        {
-            SizeChanged += OnSizeChanged;
-            MouseMove += OnMouseMove;
-            mouseMoveThrottler = new ActionThrottler(TimeSpan.FromMilliseconds(10));
-            mouseMoveThrottler.Unlocked += MouseMoveThrottlerUnlocked;
-        }
-
-        CartesianChart<SkiaDrawingContext> ICartesianChartView<SkiaDrawingContext>.Core => core;
-        public Canvas<SkiaDrawingContext> CoreCanvas => canvas.CanvasCore;
-
-        SizeF ICartesianChartView<SkiaDrawingContext>.ControlSize
-        { 
-            get
-            {
-                unchecked
-                {
-                    return new SizeF { Width = (float) canvas.ActualWidth, Height = (float) canvas.ActualHeight };
-                }
-            }
-        }
+        public CartesianChart() { }
 
         public static readonly DependencyProperty SeriesProperty =
             DependencyProperty.Register(
@@ -84,6 +51,8 @@ namespace LiveChartsCore.SkiaSharpView.WPF
             DependencyProperty.Register(
                 nameof(YAxes), typeof(IEnumerable<IAxis<SkiaDrawingContext>>),
                 typeof(CartesianChart), new PropertyMetadata(new List<IAxis<SkiaDrawingContext>> { new Axis() }));
+
+        CartesianChart<SkiaDrawingContext> ICartesianChartView<SkiaDrawingContext>.Core => (CartesianChart<SkiaDrawingContext>) core;
 
         public IEnumerable<ICartesianSeries<SkiaDrawingContext>> Series
         {
@@ -103,63 +72,12 @@ namespace LiveChartsCore.SkiaSharpView.WPF
             set { SetValue(YAxesProperty, value); }
         }
 
-        public LegendPosition LegendPosition { get; set; }
-        public LegendOrientation LegendOrientation { get; set; }
-        public FontFamily LegendFontFamily { get; set; }
-        public SolidColorBrush LegendTextColor { get; set; }
-        public double? LegendFontSize { get; set; }
-        public FontWeight? LegendFontWeight { get; set; }
-        public FontStretch? LegendFontStretch { get; set; }
-        public FontStyle? LegendFontStyle { get; set; }
-        public IChartLegend<SkiaDrawingContext> Legend => legend;
-
-        public FontFamily TooltipFontFamily { get; set; }
-        public SolidColorBrush TooltipTextColor { get; set; }
-        public double? TooltipFontSize { get; set; }
-        public FontWeight? TooltipFontWeight { get; set; }
-        public FontStretch? TooltipFontStretch { get; set; }
-        public FontStyle? TooltipFontStyle { get; set; }
-        public TooltipPosition TooltipPosition { get; set; }
-        public TooltipFindingStrategy TooltipFindingStrategy { get; set; }
-        public IChartTooltip<SkiaDrawingContext> Tooltip => tooltip;
-
-        public Margin DrawMargin { get; set; }
-
-        public TimeSpan AnimationsSpeed { get; set; } = TimeSpan.FromMilliseconds(500);
-
-        public Func<float, float> EasingFunction { get; set; } = EasingFunctions.Lineal;
-
-        public override void OnApplyTemplate()
+        public override void InitializeCore()
         {
-            base.OnApplyTemplate();
-
-            if (!(Template.FindName("canvas", this) is NaturalGeometriesCanvas canvas))
-                throw new Exception(
-                    $"{nameof(SKElement)} not found. This was probably caused because the control {nameof(CartesianChart)} template was overridden, " +
-                    $"If you override the template please add an {nameof(NaturalGeometriesCanvas)} to the template and name it 'canvas'");
-
-            this.canvas = canvas;
             core = new CartesianChart<SkiaDrawingContext>(this, canvas.CanvasCore);
             legend = Template.FindName("legend", this) as IChartLegend<SkiaDrawingContext>;
             tooltip = Template.FindName("tooltip", this) as IChartTooltip<SkiaDrawingContext>;
             core.Update();
-        }
-
-        private void OnSizeChanged(object sender, SizeChangedEventArgs e)
-        {
-            core.Update();
-        }
-
-        private void OnMouseMove(object sender, System.Windows.Input.MouseEventArgs e)
-        {
-            var p = e.GetPosition(canvas);
-            mousePosition = unchecked(new PointF((float)p.X, (float)p.Y));
-            mouseMoveThrottler.TryRun();
-        }
-
-        private void MouseMoveThrottlerUnlocked()
-        {
-            tooltip.Show(core.FindPointsNearTo(mousePosition), this);
         }
     }
 }
