+++ src/skiasharp/LiveChartsCore.SkiaSharp.WPF/Themes/Generic.xaml
@@ -4,23 +4,25 @@
     xmlns:local="clr-namespace:LiveChartsCore.SkiaSharpView.WPF"
     xmlns:skia="clr-namespace:SkiaSharp.Views.WPF;assembly=SkiaSharp.Views.WPF">
 
-    <Style TargetType="{x:Type local:CartesianChart}">
-        <Setter Property="Template">
-            <Setter.Value>
-                <ControlTemplate TargetType="{x:Type local:CartesianChart}">
-                    <DockPanel LastChildFill="true">
-                        <local:DefaultTooltip x:Name="tooltip" />
-                        <local:DefaultLegend x:Name="legend"
+    <ControlTemplate x:Key="DefaultChartTempalte" TargetType="{x:Type local:Chart}">
+        <DockPanel LastChildFill="true">
+            <local:DefaultTooltip x:Name="tooltip" />
+            <local:DefaultLegend x:Name="legend"
                                              DockPanel.Dock="{Binding ElementName=legend, Path=Dock}"
                                              Orientation="{Binding ElementName=legend, Path=Orientation}"
                                              Visibility="{Binding ElementName=legend, Path=Visibility}" />
-                        <local:NaturalGeometriesCanvas x:Name="canvas"/>
-                    </DockPanel>
-                </ControlTemplate>
-            </Setter.Value>
-        </Setter>
+            <local:NaturalGeometriesCanvas x:Name="canvas"/>
+        </DockPanel>
+    </ControlTemplate>
+
+    <Style TargetType="{x:Type local:CartesianChart}">
+        <Setter Property="Template" Value="{StaticResource DefaultChartTempalte}"></Setter>
     </Style>
 
+    <Style TargetType="{x:Type local:PieChart}">
+        <Setter Property="Template" Value="{StaticResource DefaultChartTempalte}"></Setter>
+    </Style>
+    
     <Style TargetType="{x:Type local:NaturalGeometriesCanvas}">
         <Setter Property="Template">
             <Setter.Value>
